requirements.txt
langchain-community
langgraph
apscheduler
python-dotenv 
langchain-chroma
langchain_huggingface
sentence-transformers
langchain_tavily
langchain-openai
fastapi
python-multipart